requirements.txt
discord.py
youtube_dl
pynacl
pydub
python-dotenv
moviepy
youtube_search
requests
fastapi
pydantic
pytubefix


